+++ catalog/resource_sql_table_test.go
@@ -229,11 +229,6 @@ func TestResourceSqlTableCreateTable(t *testing.T) {
 					},
 				},
 			},
-			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.bar?",
-				Response: catalog.TableInfo{},
-			},
 		}, useExistingClusterForSql...),
 		Create:   true,
 		Resource: ResourceSqlTable(),
@@ -308,11 +303,6 @@ func TestResourceSqlTableCreateTableWithOwner(t *testing.T) {
 						"three": "four",
 					},
 				},
-			},
-			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.bar?",
-				Response: catalog.TableInfo{},
 			},
 		}, useExistingClusterForSql...),
 		Create:   true,
@@ -437,11 +427,6 @@ func TestResourceSqlTableUpdateTable(t *testing.T) {
 				},
 			},
 			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.bar?",
-				Response: catalog.TableInfo{},
-			},
-			{
 				Method:   "POST",
 				Resource: "/api/2.0/clusters/start",
 				ExpectedRequest: clusters.ClusterID{
@@ -543,11 +528,6 @@ func TestResourceSqlTableUpdateTableAndOwner(t *testing.T) {
 				},
 			},
 			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.bar?",
-				Response: catalog.TableInfo{},
-			},
-			{
 				Method:   "POST",
 				Resource: "/api/2.0/clusters/start",
 				ExpectedRequest: clusters.ClusterID{
@@ -644,11 +624,6 @@ func TestResourceSqlTableUpdateTableClusterKeys(t *testing.T) {
 				},
 			},
 			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.bar?",
-				Response: catalog.TableInfo{},
-			},
-			{
 				Method:   "POST",
 				Resource: "/api/2.0/clusters/start",
 				ExpectedRequest: clusters.ClusterID{
@@ -737,11 +712,6 @@ func TestResourceSqlTableUpdateView(t *testing.T) {
 						},
 					},
 				},
-			},
-			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.bar?",
-				Response: catalog.TableInfo{},
 			},
 			{
 				Method:   "POST",
@@ -853,11 +823,6 @@ func TestResourceSqlTableUpdateView_Definition(t *testing.T) {
 				},
 			},
 			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.barview?",
-				Response: catalog.TableInfo{},
-			},
-			{
 				Method:   "POST",
 				Resource: "/api/2.0/clusters/start",
 				ExpectedRequest: clusters.ClusterID{
@@ -911,11 +876,6 @@ func TestResourceSqlTableUpdateView_Comments(t *testing.T) {
 					ViewDefinition: "SELECT * FROM main.foo.bar",
 					Comment:        "to be changed (requires new)",
 				},
-			},
-			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.barview?",
-				Response: catalog.TableInfo{},
 			},
 			{
 				Method:   "POST",
@@ -1008,11 +968,6 @@ func resourceSqlTableUpdateColumnHelper(t *testing.T, testMetaData resourceSqlTa
 					Comment:               "terraform managed",
 					ColumnInfos:           testMetaData.oldColumns,
 				},
-			},
-			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.bar?",
-				Response: catalog.TableInfo{},
 			},
 			{
 				Method:   "POST",
@@ -1402,11 +1357,6 @@ func TestResourceSqlTableCreateTable_ExistingSQLWarehouse(t *testing.T) {
 					},
 				},
 			},
-			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.bar?",
-				Response: catalog.TableInfo{},
-			},
 		},
 		Create:   true,
 		Resource: ResourceSqlTable(),
@@ -1500,11 +1450,6 @@ func TestResourceSqlTableCreateTableWithIdentityColumn_ExistingSQLWarehouse(t *t
 					},
 				},
 			},
-			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.bar?",
-				Response: catalog.TableInfo{},
-			},
 		},
 		Create:   true,
 		Resource: ResourceSqlTable(),
@@ -1571,11 +1516,6 @@ func TestResourceSqlTableReadTableWithIdentityColumn_ExistingSQLWarehouse(t *tes
 					},
 				},
 			},
-			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.bar?",
-				Response: catalog.TableInfo{},
-			},
 		},
 		ID:       "main.foo.bar",
 		Read:     true,
@@ -1584,95 +1524,6 @@ func TestResourceSqlTableReadTableWithIdentityColumn_ExistingSQLWarehouse(t *tes
 		"column.0.identity": "always",
 		"column.1.identity": "",
 		"column.2.identity": "default",
-	})
-}
-
-func TestResourceSqlTableReadTableWithPartitionColumn_ExistingSQLWarehouse(t *testing.T) {
-	qa.ResourceFixture{
-		CommandMock: func(commandStr string) common.CommandResults {
-			return common.CommandResults{
-				ResultType: "",
-				Data:       nil,
-			}
-		},
-		HCL: `
-		name               = "bar"
-		catalog_name       = "main"
-		schema_name        = "foo"
-		table_type         = "MANAGED"
-		data_source_format = "DELTA"
-		storage_location   = "abfss://container@account/somepath"
-		warehouse_id       = "existingwarehouse"
-	  
-	
-		comment = "this table is managed by terraform"
-		`,
-		Fixtures: []qa.HTTPFixture{
-			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.bar",
-				Response: SqlTableInfo{
-					Name:                  "bar",
-					CatalogName:           "main",
-					SchemaName:            "foo",
-					TableType:             "EXTERNAL",
-					DataSourceFormat:      "DELTA",
-					StorageLocation:       "s3://ext-main/foo/bar1",
-					StorageCredentialName: "somecred",
-					Comment:               "terraform managed",
-					Properties: map[string]string{
-						"one":   "two",
-						"three": "four",
-					},
-					ColumnInfos: []SqlColumnInfo{
-						{
-							Name:     "id",
-							Type:     "bigint",
-							TypeJson: "{\"type\":\"bigint\",\"nullable\":true, \"metadata\":{\"delta.identity.start\":1,\"delta.identity.allowExplicitInsert\":false}}",
-						},
-						{
-							Name:    "name",
-							Type:    "string",
-							Comment: "name of thing",
-						},
-						{
-							Name:     "number",
-							Type:     "bigint",
-							TypeJson: "{\"type\":\"bigint\",\"nullable\":true, \"metadata\":{\"delta.identity.start\":1,\"delta.identity.allowExplicitInsert\":true}}",
-						},
-					},
-				},
-			},
-			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.bar?",
-				Response: catalog.TableInfo{
-					Columns: []catalog.ColumnInfo{
-						{
-							Name:            "id",
-							PartitionIndex:  1,
-							ForceSendFields: []string{"PartitionIndex"},
-						},
-						{
-							Name:            "name",
-							PartitionIndex:  0,
-							ForceSendFields: []string{"PartitionIndex"},
-						},
-						{
-							Name: "number",
-						},
-					},
-				},
-			},
-		},
-		ID:       "main.foo.bar",
-		Read:     true,
-		Resource: ResourceSqlTable(),
-	}.ApplyAndExpectData(t, map[string]any{
-		"column.0.identity": "always",
-		"column.1.identity": "",
-		"column.2.identity": "default",
-		"partitions":        []any{"name", "id"},
 	})
 }
 
@@ -1735,11 +1586,6 @@ func TestResourceSqlTableCreateTable_OnlyManagedProperties(t *testing.T) {
 						"delta.enableDeletionVectors": "false",
 					},
 				},
-			},
-			{
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/tables/main.foo.bar?",
-				Response: catalog.TableInfo{},
 			},
 		},
 		Create:   true,