+++ internal/service/apps_tf/legacy_model.go
@@ -30,6 +30,8 @@ type App_SdkV2 struct {
 
 	AppStatus types.List `tfsdk:"app_status"`
 
+	BudgetPolicyId types.String `tfsdk:"budget_policy_id"`
+
 	ComputeStatus types.List `tfsdk:"compute_status"`
 	// The creation time of the app. Formatted timestamp in ISO 6801.
 	CreateTime types.String `tfsdk:"create_time"`
@@ -41,6 +43,8 @@ type App_SdkV2 struct {
 	DefaultSourceCodePath types.String `tfsdk:"default_source_code_path"`
 	// The description of the app.
 	Description types.String `tfsdk:"description"`
+
+	EffectiveBudgetPolicyId types.String `tfsdk:"effective_budget_policy_id"`
 	// The unique identifier of the app.
 	Id types.String `tfsdk:"id"`
 	// The name of the app. The name must contain only lowercase alphanumeric
@@ -76,12 +80,14 @@ func (c App_SdkV2) ApplySchemaCustomizations(attrs map[string]tfschema.Attribute
 	attrs["active_deployment"] = attrs["active_deployment"].(tfschema.ListNestedAttributeBuilder).AddValidator(listvalidator.SizeAtMost(1)).(tfschema.AttributeBuilder)
 	attrs["app_status"] = attrs["app_status"].SetComputed()
 	attrs["app_status"] = attrs["app_status"].(tfschema.ListNestedAttributeBuilder).AddValidator(listvalidator.SizeAtMost(1)).(tfschema.AttributeBuilder)
+	attrs["budget_policy_id"] = attrs["budget_policy_id"].SetOptional()
 	attrs["compute_status"] = attrs["compute_status"].SetComputed()
 	attrs["compute_status"] = attrs["compute_status"].(tfschema.ListNestedAttributeBuilder).AddValidator(listvalidator.SizeAtMost(1)).(tfschema.AttributeBuilder)
 	attrs["create_time"] = attrs["create_time"].SetComputed()
 	attrs["creator"] = attrs["creator"].SetComputed()
 	attrs["default_source_code_path"] = attrs["default_source_code_path"].SetComputed()
 	attrs["description"] = attrs["description"].SetOptional()
+	attrs["effective_budget_policy_id"] = attrs["effective_budget_policy_id"].SetComputed()
 	attrs["id"] = attrs["id"].SetComputed()
 	attrs["name"] = attrs["name"].SetRequired()
 	attrs["pending_deployment"] = attrs["pending_deployment"].SetComputed()
@@ -123,11 +129,13 @@ func (o App_SdkV2) ToObjectValue(ctx context.Context) basetypes.ObjectValue {
 		map[string]attr.Value{
 			"active_deployment":           o.ActiveDeployment,
 			"app_status":                  o.AppStatus,
+			"budget_policy_id":            o.BudgetPolicyId,
 			"compute_status":              o.ComputeStatus,
 			"create_time":                 o.CreateTime,
 			"creator":                     o.Creator,
 			"default_source_code_path":    o.DefaultSourceCodePath,
 			"description":                 o.Description,
+			"effective_budget_policy_id":  o.EffectiveBudgetPolicyId,
 			"id":                          o.Id,
 			"name":                        o.Name,
 			"pending_deployment":          o.PendingDeployment,
@@ -151,15 +159,17 @@ func (o App_SdkV2) Type(ctx context.Context) attr.Type {
 			"app_status": basetypes.ListType{
 				ElemType: ApplicationStatus_SdkV2{}.Type(ctx),
 			},
+			"budget_policy_id": types.StringType,
 			"compute_status": basetypes.ListType{
 				ElemType: ComputeStatus_SdkV2{}.Type(ctx),
 			},
-			"create_time":              types.StringType,
-			"creator":                  types.StringType,
-			"default_source_code_path": types.StringType,
-			"description":              types.StringType,
-			"id":                       types.StringType,
-			"name":                     types.StringType,
+			"create_time":                types.StringType,
+			"creator":                    types.StringType,
+			"default_source_code_path":   types.StringType,
+			"description":                types.StringType,
+			"effective_budget_policy_id": types.StringType,
+			"id":                         types.StringType,
+			"name":                       types.StringType,
 			"pending_deployment": basetypes.ListType{
 				ElemType: AppDeployment_SdkV2{}.Type(ctx),
 			},